segmentation_flwr/nvidia clara/requirements.txt
numpy
torch
pillow
torchvision
pycocotools
